Implement retry logic in NDSCTL
The ndsctl binary may exit with status 4 if it can't
acquire an internal lock, i.e. multiple instances are
executed concurrently.

We now recognise this condition and retry the execution
up to three times with a delay of 100ms beween attempts.

Using a simple process-local lock to prevent this
condition is not sufficient because we can't control
invocations of ndsctl made by other threads.